Make vm and mk working as single program
This commit creates a wrapper for vm and mk; this required four
actions: create a single script that wraps "vm/mk" commands, move old
scripts to a new directory name src, create a script for
install/uninstall, and rework some part of the code. First, file
easy-kernel-workflow.sh call all commands available in vm and mk.
Second, vm/mk have to be moved to a new directory; as a result, some few
changes were made in the file. Third, the "setup.sh" install the
easy-kernel-workflow as a command in the home directory and make it
available in the path. Finally, vm and mk had to be rework to extract
some common code to a file "src/common".

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>