Add routine to read local configuration
This commit adds a global configuration variable, a function for reading
configuration files, and a sample on how to create a local configuration
file. The old Qemu options were kept as a default, local configuration
file only subscribes it if the command is invoked in a directory with
the config file.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>