Add parallel compile rule to kw b
This commit adds support for parallel compilation in the "build"
command. This commit adds support for parallel compilation in the
"build" command. We define the level of parallelism by check the total
of cores available in the machine. Finally, we tried to keep this part
of the code portable for multiple distros.

Closes: #11

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>