Encapsulate all tests inside a subshell
Kworkflow unit tests previously relied on installation dependent
variables and functions for testing. This is no longer the case, as we
encapsulate all tests inside a subshell.

Function init_env in tests/utils initializes a subshell environment by
exporting a local version of kw.sh variables. This guarantees that tests
only depend on the local repository, and not the installation scripts.

Since checkpatch testing relies on some external files, run_tests.sh was
modified to check for the existence of these files in tests/external
before running any test. Also, the external files were removed from the
repository and an option 'prepare' was added to run_tests.sh, to fetch
these files, preparing the tests environment. This was done in order to
have always the newest files when testing without the need to rely on
the tester to manually update then in the repository.

Also a '--force-update' option was added to 'prepare' option, so that
the environment is reprepared, updating the external files, when
desired. README.md was modified with the information on how to use
these new run_tests.sh options.

Signed-off-by: Matheus Tavares <matheusmalavasi@uol.com.br>
Signed-off-by: Renato Lui Geh <renatogeh@gmail.com>
Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>