Add option in kw maintainers to print file authors
Now it's possible to run 'kw maintainers -a|--authors FILE' which will
print not only the maintainers of FILE but also the authors indicated in
the 'MODULE_AUTHOR' statement, inside FILE. If FILE is a directory, the
use of this option will make kw print the authors of each file under the
directory FILE (but not recursively). Note that only files in which the
'MODULE_AUTHOR' statement can be found will be printed.

Also added this new option to 'kw help' and README.md.

Signed-off-by: Matheus Tavares <matheusmalavasi@uol.com.br>