Fix run_tests.sh's test failure check
Previously run_tests.sh categorized the test files that returned 1 as
failures and all others as success. This made, for example, test files
without execution permission to be considered as successfully executed.
Now, only the ones that returns 0 are considered success and all
others are considered failures.

Also, the function report_results was printing the number of test
files that succeeded, failed and weren't found, not the number of
tests. So the strings to be printed were changed from "tests" to "test
files".

Signed-off-by: Matheus Tavares <matheusmalavasi@uol.com.br>