Add unit test for check_local_configuration
This patch adds a unit test for the check_local_configuration()
function. For the correct operation, this test relies upon two different
files: 'tests/sample/kworkflow.config' and 'tests/.tmp_commmons_test'.
The second file is a directory created and removed during the test.
Finally, the gitignore was updated to avoid the wrong inclusion of these
files.

Signed-off-by: Lucas Santos <lucas.moreira@me.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>
Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>