Add unit test for check_local_configuration
This patch adds a unit test for the check_local_configuration function
of src/commons.sh. Also added tests/sample/kworkflow.config, need for
the test, was added and the entry "tests/.tmp_commmons_test" was added
to .gitignore. This is a directory created and removed during the test.
The .gitignore entry was added as a precaution.

Signed-off-by: Lucas Santos <lucas.moreira@me.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>