Remove `source` from kw.sh
This commit removes the dependency on sourcing kw.sh. Although this
commit does not explicitly add support to other shells, such as ksh and
fish, it makes kw callable from other shells. As a proof of concept, a
new test was added with the goal of call kw from dash.

It also updates the README.md with new information regarding dependence
packages, and how to install `kw` with other shells.

Signed-off-by: Giuliano Belinassi <giuliano.belinassi@usp.br>