Fix the extra '=*' in the codestyle command
Currently, if we use 'kw c' in a source file without any warning, it
displays something like this:

$ kw c drivers/gpu/drm/vkms/
Running checkpatch.pl on: drivers/gpu/drm/vkms/

=========================================================
PATH:70: ANY OUTPUT
total: 0 errors, 0 warnings, 1 checks, 144 lines checked

=========================================================

=========================================================

=========================================================

=========================================================

This patch, fix this problem by removing the extra "=*". Additionally,
this commit updates the test to better fit this new output and add a new
check for the case that does not have any issue in the code.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>