Update explore_test to get only the first match
The explore_test.sh have a particular case that checks for the word GNU
in the LICENSE file. After we changed the LICENSE from GPLv3 to GPLv2
this test started to break, because GPLv2 has a double match in the test
case instead of one. With this idea in mind, this commit adds the '-m'
flag to the grep command with match limit of one.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>