src: mk: Add check for the case that mount fails
If the vm is running and the user tries to install a new version of a
module (kw bi), the install operations continue. This is a problem
because the module is installed at ~/p/mount/lib instead of
QEMU_IMAGE/lib. This commit aborts the install operation in this case.

Closes: #1

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>