src: mk: use ${configurations[arch]} instead of $MAKE_OPTS
We currently pass the $MAKE_OPTS variable to 'make' when 'kw build' is
invoked. However, this variable is outdated as it is never defined. In
addition, we now have a ${configurations[arch]} user setting which
defines the arch to build but we don't currently pass it to make. So,
solving both problems at once, remove the outdated $MAKE_OPTS and invoke
make with the architecture option instead.

Fixes #63

Signed-off-by: Rodrigo Ribeiro Carvalho <rodrigorsdc@gmail.com>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>