documentation: contribution: remove email-based instruction
The Git documentation (which we borrowed the D-C-O text from), says to
add a "From: " header in the patch to identify the true author. This
instruction is to be used on email-based workflows, where the patch is
sent my email. With GitHub or Gitlab, however, the author is already
identified in the Pull Request, so this instruction is not needed and
may confuse contributors. So let's remove it.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>