tests: checkpatch: fix test condition which always eval to success
kw learned to use the local checkpatch.pl script instead of a downloaded
file at installation, but we forgot to adjust the tests. Despite the new
behavior, most tests are still running successfully (even though they
should be failing). This problem is occurring due to a wrong argument
usage in assertTrue which is making it always evaluate as success. Let's
fix that, leading to the expected test failure, which will be fixed in
the following patch.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>