tests: checkpatch: fix no codestyle error test case
We currently use 'testNothing' to test `kw c`'s output for a file that
doesn't contain codestyle errors. But there're some bugs limiting the
efficiency of this test.

The test is currently failing because there's a typo 'assertFail' ->
'assertFalse'. Fixing this would make the test succeed, however, it
still wouldn't be testing what we need in this situation, since the
given path is invalid. (It will always succeed because the output will
be 'Invalid path', and we only test if the output does not contain
'total')

So, instead, let's remove this case and reintroduce the 'testCorrect'
case, added in bd93d69, which looks for a "=*" string in kw's output.
This is still not the desired behavior, as kw should not print anything
when no codestyle error is found. But this is what kw is currently
doing, so let's adjust the test to check this string (at least for
now).

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>