Implement system-wide installation
This patch add a Makefile with a few targets in order to add
a system-wide installation.

Because of the new install destination, we modify several files
to ensure a correct files paths and correctness of tests.

Now every call of 'kw' will trigger a verification whether user's HOME
has a .config/kw directory. If not, that dir will be create with
kworkflow.config file.

Signed-off-by: Rodrigo Carvalho <rodrigorsdc@gmail.com>