Remove local config file example
The `kworkflow.config.example` file is an example of what a local
configuration file may contain. But it is outdated and no longer
represents everything users can customize. So, let's remove this file
and ajust the respective docs section that referenced it to talk about
`etc/kworkflow.config`, instead. Besides being a more complete example
of all possible settings, it is the default configuration file, thus it
is much less likely to become outdated.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>