docs: Update docs and commands due to deploy feature
After we added the deployment feature, we shifted kw focus from QEMU to
a more generic targets; more precisely, any target machine. For this
reason, this commit updates the documentation for a better portrait of
the kw usage. Additionally, we renamed and dropped some commands that do
not make sense for this new context.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>