Drop prepare feature
When we started to develop kw, we want to make the Kernel developer's
life as easy as possible. For this reason, we added a feature named
prepare that uses Ansible scripts to put a VM in a well-known state.
However, after a long time of experiments, we realize that it is hard to
maintain such feature generic, and it also adds much complexity.
Additionally, some people start to package kw for Arch and Debian, which
means that Ansible represents another dependency for kworkflow. For all
of these reasons, we decided to drop this feature because it adds much
complexity, and the final result does not bring many benefits for
end-users.

Closes #138

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>