WIP: make installation closer to unix standards
This patch makes kw files be installed in the most correct places
according to unix standards. A new --prefix= option is also added to
setup.sh to allow localized installations.

WIP note: This is a *major* change to the codebase, so we must make sure
everything works before merging. Currently, the patch is marked as WIP
because it is a *very* early vertion without testing and probably with
many bugs. I would suggest reviewers to start reading by `setup.sh` and
`kw` to understand the rationale behind the changes. Finally, the
`tests` directory hasn't been touched yet so, without adjustments, it
will fail for now!

Solves #50 and #136

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>