kwlib: Rework get_from_colon to make it generic
We have the function get_from_colon() which is intended to handle
strings that have ':' character, this is useful for handling IP:PORT
format. However, the nature of this function could be easily promoted to
something more generic in order to make it more powerful; for this
reason, this commit rename get_from_colon() to get_based_on_delimiter()
and adds a new parameter that indicates the character delimiter that
should be used. Rework this function also solve the issue of handling
hostname in the parameter `ssh_ip` on `kworkflow.config`, after this
change users can finally use `ssh_ip=username@hotname`.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>