src: mk: Add kernel list support to deploy option
kw deploy already support the installation of new kernels on remote,
local, and virtual machine (partially); however, it still lacks a way to
show all installed kernel and a mechanism to uninstall a single or a
group of kernels. This commit adds support for listing all installed
kernels in a target machine by adding a new options to list available
kernels, users can use it with these commands:

 kw deploy [--vm|--local|--remote] --ls [--ls-line]

The `ls-line` option shows all available kernels in as a list separated
by ','; this is a preparation for the uninstall feature. Finally, this
commit had to change the script `run_test` for enabling plugin tests.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>