setup: make installation closer to unix standards
This commit makes kw files be installed in the most correct places
according to UNIX standards. A new --prefix= option is also added to
setup.sh to allow localized installations. Unfortunately, this is a
complicated change that requires touching multiple files; for this
reason, this commit updates some functionalities behavior. In summary,
this path update requires updates in the following areas:

* Update path variables names;
* Adapting validations based on new paths;
* Drop support to fish because we nobody is using it;
* Update templates files.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>