src: plugins: kernel_install: utils: Remove code duplication
Installing a new kernel in Arch and Debian-based systems is very
similar; the major difference is how each distro generates its temporary
root filesystem. After this observation, this commit moved all the
specific code to a single place and extracted the mechanism for
generating the temporary root file system to a distro-specific file.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>