help: Break help function by sub-commands
A single help command for the entirety of kw does not scale, it becomes to long and complexe to be useful.
By keeping only a short description of each subcommand as the main kw help,
we can distribute the details and options of each of them as their own help function.
This PR implements this by breaking down the main help function into auxiliary help functions for each subcommand.

This PR also adds missing entries to the main kw help and order them alphabetically.
closes #180

Signed-off-by: alan-barzilay <alan.barzilay@gmail.com>