src: kwio: Add error message for wrong commands set in the alert variables
This commit instructs users on how to proceed in case they assigned a
command that doesn't exist to either one of the alert variables in the
kworkflow.config file.
To do so, a new function called command_exists was added, as well as a
new test related to the sound alert.

Closes #219

Signed-off-by: Magali Lemes <magalilemes00@gmail.com>