src: kw_time_and_date: Add date and time helper functions
In the kw code, we are using date and time manipulation in the statistic
code, and we are planning to add other features which will benefit from
a sound date/time management system. This commit introduces a new set of
functions responsible for dealing with date and time; as a way to
demonstrate how to use it, this change also updates the file statistics
to leverage this library.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>