src: statistics: Fix year option
The year variable shouldn't have had a month included. By removing
the month from the variable year, the 'kw statistics --year' command
works when a user specifies a year.

A test validating the --year option was also added, in which we test
two different years. To achieve that, a new folder named 2021 was
added to the test samples directory.

Closes #254

Signed-off-by: Magali Lemes <magalilemes00@gmail.com>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>