src: pomodoro: Add list label option
When users are working with Pomodoro in a way that they want to generate
reports, they can use the tags feature. However, we do not provide a
mechanism for users to check their previously registered tags which can
be challenging for them to reuse tags and generate a meaningful report.
This commit fixes this problem by adding support to list all tags
previously registered.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>