src: signal_manager: Implement signal manager
Currently, kw does not handle signals. If you pass it a SIGINT, for
example, the subprocess that happens to be executing at that particular
moment is the one who will receive and treat it. This results in a poor
user interface, where the standard Ctrl-C doesn’t have a meaningful
outcome. Worse than that, a simple SIGINT could leave the environment
polluted by unremoved temporary files. This patch:

 - Implements functions for adding and resetting handlers
 - Implements default handler for SIGINT and SIGTERM
 - Describes a use policy of signal handling in
   documentation/content/howtocontribute.rst

closes: #270

Signed-off-by: João Seckler <jseckler@riseup.net>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>