src: remote: stop using root user
Doing a ssh operation as root is a major security risk, one that our
end-users may not want or even be able to take. This commit refactors
the functions cmd_remotely and cphost2remote to use sudo instead of the
root user.
drm: adapt drm and its tests for non root ssh

Signed-off-by: Alan Barzilay <alan.barzilay@gmail.com>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>