src: deploy: Simplify how we get IP and Port
We have a complex way of handling IP and port, which makes things a
little bit hard to improve, adds unnecessary complexity to something
simple, and plasters the deploy expansion. This commit eliminates all of
these complexities by adding two new entrances to "options_values":
REMOTE_IP and REMOTE_PORT. Instead of calling a function using a
specific format, we can just access this information directly from the
data struct; however, we expect that the parser option deals with all
the complexity around the IP and port manipulation.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Reviewed-by: Alan Barzilay <alan.barzilay@gmail.com>
Reviewed-by: João Seckler <jseckler@riseup.net>