src: kw_include: Disambiguate files with same name
Previously, files with the same name but in different folders couldn't
be included concomitantly, because the include manager would use the
same variable to keep track of which file was already included. This
commit introduces a method for creating such variable that is based on
the path of the file, not only its name.

Signed-off-by: João Seckler <jseckler@riseup.net>
Reviewed-by: Alan Barzilay <alan.barzilay@gmail.com>