src: remote: update default values logic
As previously discussed we should leave the handling of default values
to the kw_config_loader file. There is no need to re-establish the same
logic present in our hierarchical parsing of config files. This commit
removes duplicated code and simplifies some functions of the remote.sh
file.
This commit also updates the  config template present in the
samples folder as well as remove the name restriction of config file
names.

Signed-off-by: Alan Barzilay <alan.barzilay@gmail.com>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>