src: deploy: Refactor option parsing
Use getopt to simplify the parsing of options for kw deploy. This
implies one change of behaviour: getopt accepts abbreviations of
options, as long as there is no ambiguity. For example, now --rem is
accepted and is equivalent to --remote. Also, the --remote option for kw
deploy now requires an option.

Signed-off-by: João Seckler <jseckler@riseup.net>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>