src: kw_ssh: Create a file dedicated to handling ssh connections
The code related to ssh was moved to the remote file; however, this is
incorrect because 'remote.sh' should work as a library. This commit
creates a new file named 'kw_ssh.sh' and 'test/kw_ssh_test.sh' and moves
all the ssh functions from 'remote.sh' to these files. This commit did
not change the logic; it just moved the code and made the necessary
adaptations.

Closes #395

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Reviewed-by: Magali Lemes <magalilemes00@gmail.com>