Make kw compliant with XDG Base Directory Specification
User customizations for kw folder should now be done with the XDG_*_HOME
environmental variables, instead of PREFIX. Also, ~/.kw is no longer
recognized as a path for a configuration file (unless it is supplied in
XDG_CONFIG_DIRS). This also fixes some problems with DEV-MODE: corrects
the paths to documentation, cache and data folders.

Closes: #296 and #385

Signed-off-by: João Seckler <jseckler@riseup.net>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>