src: remote: Update code to correctly uses remote_parameters
I made a patch that refactors how we use remote info to enable users to
specify a user name for ssh command. Unfortunately, I did not replace
the configurations[ssh_*] to remote_parameters[*] everywhere, and some
inconsistent results appeared. This commit addresses some of these
issues and also updates the remote test to this new context.

Fixes: e5e8d37 ("src: remote: Get user as a parameter")
Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>