explore: fix explore help text
The help text for the explore command was a bit confusing, there was a
weird parameter. I realized its purpose was to show that strings with
spaces need to be quoted. I don't think this is needed in the help text
as it is clearly stated in the man page.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>