src: kw_config_loader: Remove extra space from variable
When I was changing my config file, I accidentally forgot an extra space
at the end of the IP address; this simple mistake leads to tons of
failures because when we compose the IP with other commands, kw also
adds the extra space. This commit fixes this issue by adding a simple
pos-data process to remove any extra space from the beginning and the
end of the variable.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>