setup: Fix legacy_folders condition
We recently updated kw to make it compliant with XDG, and we added a
function to migrate the old kw config to the new model automatically.
However, one of the conditions is always false, and for this reason, we
could not update the file location for features like configm and
pomodoro. This commit fixes this issue by handling the migration case by
case.

Fixes: e30c0a86 ("Make kw compliant with XDG Base Directory Specification")
Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Reviewed-by: Magali Lemes <magalilemes00@gmail.com>