src: pomodoro: check if POMODORO_LOG_FILE exists
In the function show_active_pomodoro_timebox, we weren't testing
if POMODORO_LOG_FILE existed before reading it, which leads to an
error when invoking 'kw p -l' without this file in the system. This
commit adds a condition to test the existence of POMODORO_LOG_FILE
before attempting an operation with it.

Closes #412

Signed-off-by: Magali Lemes <magalilemes00@gmail.com>