src: time_and_date: remove cal dependency
When we introduced the Pomodoro version, we add a dependency to the
`ncal` package; additionally, in the latest version of shunit we noticed
that this approach made our shunit test a little bit unstable. In an
effort to keep kw as minimal as possible and make the unit test stable,
this commit drops the use of cal.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>
Acked-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>