src: deploy: Simplify prepare remote function
The function prepare_remote_dir has multiple steps that require opening
and closing three connections, which is a waste of time and increases
the unit test complexity. This commit simplifies all of the preparation
steps to a single command, reducing the number of connections and
simplifying the deploy test.

Changes since V1 (Rubens):
- Grammar improvements
- Use single-quote instead of double-quote

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>