src: device: Change how we approach the DE
When we try to identify the desktop environment, we look at the session
folder, which is not reliable. For trying to make the output more
reliable, this commit checks for the potential process currently
running. Keep in mind that this is not 100% reliable, but we don't have
any easy way to deal with this problem.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>