src: configm: Rework fetch operations
This commit is a code refactor on how we get a config file from a target
machine. Previously most of the code was centralized in a single
function; after this rework, we moved specific parts to their own
function and expanded the unit test to better cover this new scneario.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>