src: plugins: kernel_install: Avoid duplication in the list
If we deploy the same kernel multiple times, we keep adding the
duplication in the list, which is not ideal. This commit fixes this by
adding a simple check before writing anything to the file in the remote;
additionally, we make sure that we deploy the latest script to make sure
that kw uses the newest version.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>