src: vm: unify mount, umount and up under vm option
The mount, umount and up functions are one of the first functions
provided by the KernelWorkflow, so, for this reason, they are
currently separated in three diferent flags, even though they are
all related to the same operation: managing a QEMU VM. Solves #453.

Creates a general vm function, that will be used to redirect to the
original functions mount, umount and up, making all the changes
related to it, as the documentation and how the functions are
called in the tests. So, it gives more coherence to the
KernelWorkflow code and improves organization.

Co-authored-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>
Co-authored-by: Paulo Guilherme <pauloguilherme@usp.br>