tests: kw_time_and_date_test: fix bug
The function test_get_today_info sometimes failed when a second passed
between calling date and get_today_info functions. Solves #308

Adds a new variable to check the date return before and after the
calling of the get_today_info function. Checks if they are the same
and, if they are not, repeat the process. Assuring that both calls
were made in the same second.

It was also considered to use a while loop instead of an if, but
to avoid problems with possible endless loops, this solution
was disregarded.

Signed-off-by: Paulo Guilherme <pauloguilherme@usp.br>
Co-authored-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>