tests: kw_time_and_date_test: fix bug
The function test_get_today_info sometimes failed when a second passed
between calling `date` and `get_today_info` functions. Solves #308

Check if `today` and `formatted_today` are the same and, if they
are not, test them again. Making sure that both calls were made in the
same second if they were not before. Correct grammar error in
`formatted_today`.

It was also considered to create a new variable to double-check `today`
value. Another possible idea was to use a while loop instead of an if,
but to avoid problems with possible endless loops, this solution
was disregarded.

Signed-off-by: Paulo Guilherme <pauloguilherme@usp.br>
Co-authored-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>