setup: add kw man page to mandb
Adds a compressed copy of the main man page of kw to a location that the
man command can find so that running 'man kw' becomes possible.

Also stops synchronization of raw documentation files as they are not
needed for any features to work, so no need to clutter the users system.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>