src: kw_include: fix variable naming
In some systems, kw couldn't deal properly with hidden folders
(e.g. #473), as the variables created for naming includes had dots on
them, which made them improper (i.e. the user would see 'not a valid
identifier' errors popping up for every `kw` call).

This commit swaps both dots and spaces for underscores in those file
paths.

Also, as the include command returns 0 in spite of the declare having
failed, check the declare retval output to make sure it is working as
expected.

Signed-off-by: Isabella Basso <isabbasso@riseup.net>