src: init: Interactive mode for SSH
Adds a new function to init.sh, called
interactive_init, which is responsible for handling
init.sh in interactive mode. Also adds a text file to
src/strings/init.txt which keeps most of the text
used in interaction. Modifies init.sh to use
interactive_init and the ssh lookup functions.
Functionalities yet to be implemented for this goal
are commented with #TODO in the body of the code.

Co-authored-by: Miguel de Mello Carpi <miguelmello@usp.br>
Co-authored-by: Lourenço Henrique Moinheiro Martins Sborz Bogo <louhmmsb@usp.br>
Signed-off-by: Eduardo Brancher Urenha <eduardo.urenha@usp.br>