Add interaction support to init.sh for ssh setup and remote config
In this commit we modified init.sh to look for the --interactive
(or -i) options when parsing arguments. If these options are found,
the script will attempt to interactively detect and setup SSH in
the user's system. Prompts are added to allow the user to interact
with this process and skip steps if desired. We also added an
interactive configuration step for the user's remote connection
settings after SSH is setup. These functionalities are intended to
provide a better user experience for new users who may never have
setup KW before. Functionalities yet to be implemented for this
goal are commented with a #TODO: in the body of the script.

Co-authored-by: Miguel de Mello Carpi <miguelmello@usp.br>
Co-authored-by: Lourenço Henrique Moinheiro Martins Sborz Bogo <louhmmsb@usp.br>
Co-authored-by: Eduardo Brancher Urenha <eduardo.urenha@usp.br>