src: kwio: add ask_with_default feature
Currently, there is no easy way to ask the user information with a
default option (something that makes sense to the user), so the user
needs only to type ENTER to use that suggestion.

Add a feature that generalizes this process, with a message and the
default option being given as parameters. There is also the possibility
to send optional parameters, the third possible argument is to select if
the default option should be printed to the screen (allowing the
programmer to format it in the way the situation asks more precisely),
and the fourth parameter is a flag used to test this function.

Reviewed-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>
Co-authored-by: Paulo Guilherme <pauloguilherme@usp.br>
Signed-off-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>