src: plugins: kernel_install: Add helper to detect partition type
If the target system uses MBR, there is a good chance that the first 4
bytes inform the partition table type. For this reason, this commit
introduces a function that checks those bytes. It is important to
highlight that we also added a couple of binary files to the sample
folder for validating the new function and other news that will be
implemented.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>