src: mail: hide user pass on --list
Hides the user mail password as it is a potentially sensitive
information that should not be carelessly displayed in plain text.
This solution is not particularly safe and immune to attack, but it
access information that is stored in plain text on the computer, so it
shouldn't be a problem.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>