src: mail: deal with empty options
There are setups that may require an option to be set to an empty field.
This would cause kw to treat the option as not set and skip confirmation
before changing the value, which could cause unwanted changes to the
user system. It also caused the `--list` option to not show these
values, now when this situation is encountered an `<empty>` placeholder
value is used to mark the option.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>