src: mail: refactor mail_setup
Incorporates the functionality of check_add_config into the mail_setup
function, this recontextualizes the latter function for the eventual
implementation of the template and interactive functionalities.
This also polishes some of the messages and clarifies the changes being
done.
Applying the changes is now the last thing the setup does, this way
it reduces the chances of an having an incomplete setup.

Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>