src: kwio: Adds load_module_text to kwio.sh
load_module_text is a function that creates a dictionary from a
.txt file following a particular syntax, as follows:
[key]:
text
The created dictionary is accessed by ${module_text_dictionary[key]}.
This is useful to store long strings outside of the code.
Adds a little test suite for this function in kwio_test,
and a couple test samples in tests/samples.

Co-authored-by: Miguel de Mello Carpi <miguelmello@usp.br>
Co-authored-by: Lourenço Henrique Moinheiro Martins Sborz Bogo <louhmmsb@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Eduardo Brancher Urenha <eduardo.urenha@usp.br>