src: db: introduce db interface
In an effort to standardize our data collection and storage, as well as
improve scalability this starts a library to create and interact
with `sqlite3` databases. This will eventually replace the files and
folders found in `KW_DATA_DIR` with a unified database.
This first commit creates a function `execute_sql` that executes a SQL
script on a given database.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>