src: report: resolve compatibility of statistics with output option
Refactors the output of report, adding the --pomodoro option that covers
the output previously shown by report, as well as the --all option that
shows both --pomodoro and --statistics in that order. Also, the output
logic is refactored to ensure compatibility between all outputs
generated and the possibility of using --output.

Signed-off-by: Caique de Camargo<caique.de.camargo@hotmail.com>
Signed-off-by: Enzo Shiraishi<enzo.shiraishi@gmail.com>
Signed-off-by: Lucas Theodoro Guimaraes de Almeida <lucastheo.g.a@gmail.com>