src: debug: Ensure that we always disable ftrace
While I was using the debug feature, I noticed that the switch between
ftrace and event is incorrect because we need to set nop to the
current_trace file; otherwise, ftrace will have higher precedence over
events. This commit fixes this issue by always setting nop to
current_trace at the end of the debug.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>