src: debug: Add reset feature
When I was playing with this feature, I found myself in a situation
where we left some file set from the previous debug, which caused
problems when I tried to utilize this feature again. This commit
introduces a reset feature responsible for setting all debug files to
their default and killing any hanging process in the trace_pipe file to
address this problem.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>