tests: utils: fix bug in mk_fake_git
There were cases where the function `mk_fake_git` resulted in errors
due to not being confident if the name and email were configured.
To be sure that they are correctly configured, it was necessary to
change their order in the function. Also, add documentation to the
function and update the `printf` format.

Signed-off-by: Paulo Guilherme <pauloguilherme@usp.br>
Co-authored-by: Lucas Cardoso  <lucas.cardoso.santos@usp.br>