src: db: introduce db model and migration script
Introduces the file that builds the model to be used by kw to store all
data related to statistics options, pomodoro sessions and kernel config
files.

Adds a script to handle the data migration from the metadata files in
`KW_DATA_DIR` to the newly introduced database. Current migrated data
are: statistics and pomodoro reports, and kernel configs. The
directories holding the metadata are renamed as such:
- `KW_DATA_DIR`/statistics->`KW_DATA_DIR`/legacy_statistics
- `KW_DATA_DIR`/pomodoro->`KW_DATA_DIR`/legacy_pomodoro
- `KW_DATA_DIR`/configs/configs->`KW_DATA_DIR`/configs/legacy_configs
- `KW_DATA_DIR`/configs/metadata->`KW_DATA_DIR`/configs/legacy_metadata

Thanks: Magali Lemes <magalilemes00@gmail.com> for the help modelling
  the database.
Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>