src: deploy: Update deploy paths
This commit is preparation work to make the deploy feature slightly more
generic by using better paths. Additionally, it moves those paths to a
set of variables that can be easily updated, enabling us to give the
user the option to set up the path they want in the config file.

Closes: #470.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>