src: plugis: kernel_install: Fix vm deploy
While conducting this massive code refactor, I took the strategy of
first focusing on local and remote deploy even if it means breaking vm
support. This commit represents a fix to the vm deploy, which means that
kw vm is working again for Debian and Arch VM. Note that we still need
to bring much more improvements to the VM parts; this commit is far away
from being ideal for vm deploy.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>