src: kwlib: potential bug?
Upon uploading a change our CI unit tests failed, I couldn't reproduce
the failure locally and a re-run of the tests was successful. I tried to
follow the excution flow to figure out what could have caused this
problem and this line is the most likely candidate. This commit is for
discussion purposes only.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>