src: deploy: Enable path configuration for remote deploy
Currently, kw does not provide an easy mechanism to change where kw
deploys the necessary scripts and the temporary folder, which can add
unnecessary constraints to users. This commit introduces two new
configuration variables that enable users to specify where they want to
save kw files in the remote and which folder they want to use for
temporary files during the deploy.

Closes: #470

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>