src: deploy: Reduce code duplication
When we run kw remote, we have to deal with a tiny interface that
expects two parameters every time; as a result, in the deploy code, we
repeat those standard parameters over and over again. This hurts the
code readability and damages the idea of interface; this commit solves
this problem by creating a standard prefix to interact with the remote
interface.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>