src: init: create interactive init feature
In an attempt to make KW more user-friendly, this PR adds a new
feature to help beginners users to configure what is needed to
work with it.  To do so, it is necessary to set some tools that
are commonly used by developers (such as git).

The idea is to explain what are those configurations and suggest
default options helping beginners to set their environment.
This commit does it with git.

We selected to configure user name, email, editor, and branch name
since these are the configurations in this tutorial:
https://git-scm.com/book/en/v2/Getting-Started-First-Time-Git-Setup.

Signed-off-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>
Signed-off-by: Paulo Guilherme Pinheiro <pauloguilherme@usp.br>
Co-authored-by: Eduardo Brancher Urenha eduardo.urenha@usp.br
Co-authored-by: Miguel de Mello Carpi miguelmello@usp.br
Co-authored-by: Lourenço Henrique Moinheiro Martins louhmmsb@usp.br