src: kwlib: Use an array to handle the distro family
We are expanding the set of distros supported by kw; as a result, the
detect_distro was implemented in a way that is hard to expand the set of
the supported distro. This commit adds an array per distro family to
address this problem, where each element represents a specific
derivative distro.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>