src: deploy: Create pack function for kernel files
When deploying a kernel to a target machine, we have multiple copy steps
spread in different parts of the code, which is a problem since it makes
kw deploy code every trick. This commit creates a function responsible
for packing all the required files for the boot folder in a single place
and send to the target device. This change represents a code refactor in
the deploy, which required multiple changes in the test. Since various
tests broke during the development of this feature, I also took this
opportunity to refactor our tests to make them more manageable.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>